SOAP message mapping improvments -- use the WSDL soap:header soap:body info.
Changed API related to fault reporting to use a QName instead of String.
Tweaked OReply to contain fault QNAME not WSDL4j Fault structure. 



git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/ode/trunk@504788 13f79535-47bb-0310-9956-ffa450edef68

+++ bpel-runtime/src/main/java/org/apache/ode/bpel/engine/BpelServerImpl.java
@@ -307,7 +307,7 @@ public class BpelServerImpl implements BpelServer, Scheduler.JobProcessor, Proce
             }
 
             if (create) {
-                ProcessDAO newDao = conn.createProcess(pid, oprocess.getQName(), oprocess.guid, version);
+                ProcessDAO newDao = conn.createProcess(pid, oprocess.getQName(), oprocess.guid, (int)version);
                 for (String correlator : oprocess.getCorrelators()) {
                     newDao.addCorrelator(correlator);
                 }
@@ -390,7 +390,7 @@ public class BpelServerImpl implements BpelServer, Scheduler.JobProcessor, Proce
                 while (true) {
                     Thread.sleep(pollingTime);
                     _mngmtLock.writeLock().lock();
-                    try {
+                    try { 
                         __log.debug("Kicking reaper, OProcess instances: " + OProcess.instanceCount);
                         // Copying the runnning process list to avoid synchronization
                         // problems and a potential mess if a policy modifies the list
